feat(middleware): Refactor keyspace checker for kernel types and improve forwarding test
- The `KeyspaceCheckerMiddleware` now differentiates between classic and nextgen kernel types for keyspace validation.
- In classic mode, keyspace checks are bypassed, and an empty keyspace meta is set.
- In nextgen mode, a keyspace query parameter is now mandatory and must be an `ENABLED` keyspace.
- `TestKeyspaceCheckerMiddleware` has been split into `_classic_test.go` and `_nextgen_test.go` for build tag-specific testing.
- The `TestForwardToCoordinator` has been improved to use a real `httptest.Server` for more realistic forwarding simulation.
